pyproject.toml
[project]
name = "orange-hrm-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "allure-pytest>=2.15.3",
    "faker>=40.12.0",
    "playwright>=1.58.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.5",
]
[tool.pytest.ini_options]
addopts = [
"--tracing=on"
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
